requirements.txt
Flask==2.1.2
python-dotenv==0.20.0
requests==2.27.1
google-generativeai==0.5.2
IPython==8.24.0
Pillow==9.3.0